Abstract away the details of SpiderMonkey's debugging mode.
Added a new interface GjsInterruptRegister. This is a central register
which will ideally be provided by Gjs for internal and external modules to
insert debugger breakpoints, set single-step mode on and get interrupts
from SpiderMonkey without exposing the details of SpiderMonkey itself.

GjsInterruptRegister is an abstract interface describing an "interrupt
register". Clients can use the add_breakpoint, set_single_step_mode etc
functions to register relevant callbacks. It will return a GjsDebugConnection,
which is an object that holds a callback internally used to unregister
the breakpoint, single-step-mode etc on destruction. All that clients need to
do in order to remove their debug hook is unref this object.

GjsDebugInterruptRegister is an implementation of GjsInterruptRegister that
actually manipulates the SpiderMonkey debugging mode.

The way it generally works is to toggle certain modes on or off by using a
refcount to do so. Whenever the callback function is called for a relevant
debugging hook it dispatches the result to the relevant callback function
that was registered.

This enables code coverage tools, debuggers, profilers, etc to all co-exist
without fighting over the debugging mode. It also avoids leaking JS
engine implementation details into clients.